+++ src/main/java/codejava/Responsitory/StatsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import codejava.Entity.Orders;
 @Repository
-public interface StatsRepository extends JpaRepository<Orders, Long>{
+public interface StatsRepository extends JpaRepository<Orders, Integer>{
 	@Query(value = "{CALL sp_getTotalPriceOneMonth(:month, :year)}", nativeQuery = true)
 	String getTotalPriceByMonth(@Param("month") String month, 
 			@Param("year") String year);
